change behaviour of S0 response assignment in GTM
This commit changes the way in which the S0 of every tissue is passed to
the model constructors. Now it is allowed to pass None, if the
corresponding model has been defined using the `S0_tissue_responses`
parameter. All of this happens in the `MultiCompartmentModel.fit`
function.

Other changes involve the proper definition of the tortuosity constraint
that have been addressed in the parent commits.