FIX: pass scheme to call of MultiCompartmentSphericalHarmonicsModel
This commit changes the behaviour of the `simulate_signal` and the `__call__` function of instances of the
MultiCompartmentSphericalHarmonicsModel class by changing the
`_construct_convolution_kernel` method.

Before this commit, the acquisition scheme employed by the method was
the one contained in the `self.scheme` attribute. After applying this
commit, the employed acquisition scheme is the one passed as the keyword
argument `acquisition_scheme=` to the method. If it is not passed, the
function will use the one contained in `self.scheme` for backward
compatibility.