transform T1_tortuosity into class
This commit is responsible for transforming the T1_tortuosity function
into a class. This is made necessary by the introduction of the
correction of tortuosity constraint in generalized tissue modelling.
The corresponding changes in the constraint checks are applied both in
the modeling_framework file and in the distribute_models file.

Also, this commit solves several pep8 and other minor style issues.