add directions to input of forward_model_matrix in AMICO models
This commit changes the signature of the fit_func of the
MultiCompartmentAMICOModel class, which now requires the voxel-specific
directions to be passed.

The check for orientations to be fixed a priori has been fixed to take
into account when distinct orientations have been fixed for each voxel.

Still to be added:
- Check if some non-directional parameter have been set to
  voxel-specific values (forbidden)
- Create a fitted_parameters volume of the right dimension